reorganize pca funcs
