+++ dynamo/tools/clustering.py
@@ -14,7 +14,7 @@ from ..configuration import DKM
 from ..dynamo_logger import main_info
 from ..preprocessing.preprocessor_utils import log1p, normalize
 from ..preprocessing.QC import filter_genes_by_outliers as filter_genes
-from ..preprocessing.utils import pca
+from ..preprocessing.pca import pca
 from ..utils import LoggerManager, copy_adata
 from .connectivity import _gen_neighbor_keys, neighbors
 from .utils import update_dict
